requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
openai==1.3.0  # оставляем как резервный вариант
pyTelegramBotAPI==4.*   # telebot
yandex-speechkit
requests
easyocr
SpeechRecognition
pydub
ffmpeg-python
numpy
Pillow